@@ -8,20 +8,24 @@
   "homepage": "",
   "repository": "",
   "scripts": {
-    "dev": "vite build --watch",
+    "dev": "cross-env NODE_ENV=development VITE_SOURCEMAP=inline vite build --watch",
     "release": "node ./release.js",
     "release:manual": "node ./release.js --mode=manual",
     "release:patch": "node ./release.js --mode=patch",
     "release:minor": "node ./release.js --mode=minor",
     "release:major": "node ./release.js --mode=major",
-    "build": "vite build"
+    "build": "cross-env NODE_ENV=production vite build",
+    "make_dev_copy": "node --no-warnings ./scripts/make_dev_copy.js"
   },
   "dependencies": {
     "simple-mind-map": "^0.11.2"
   },
   "devDependencies": {
+    "@sveltejs/vite-plugin-svelte": "^3.1.0",
+    "@tsconfig/svelte": "^4.0.1",
     "@antfu/eslint-config": "^4.8.1",
     "@types/node": "^20.2.0",
+    "cross-env": "^7.0.3",
     "@vue/tsconfig": "^0.7.0",
     "eslint": "^9.22.0",
     "eslint-plugin-format": "^1.0.1",
@@ -31,11 +35,12 @@
     "minimist": "^1.2.8",
     "rollup-plugin-livereload": "^2.0.5",
     "sass": "^1.62.1",
-    "siyuan": "1.1.5",
+    "siyuan": "1.1.6",
+    "svelte": "^4.2.19",
     "ts-node": "^10.9.1",
-    "typescript": "^5.0.4",
-    "vite": "^6.2.1",
-    "vite-plugin-static-copy": "^0.15.0",
+    "typescript": "^5.1.3",
+    "vite": "^5.2.9",
+    "vite-plugin-static-copy": "^1.0.2",
     "vite-plugin-zip-pack": "^1.0.5"
   },
   "pnpm": {
