@@ -1,162 +1,180 @@
-/* eslint-disable node/prefer-global/process */
-import { resolve } from "node:path"
-import fg from "fast-glob"
-import minimist from "minimist"
-import livereload from "rollup-plugin-livereload"
-import {
-  defineConfig,
-  loadEnv,
-} from "vite"
+import { resolve } from "path"
+import { defineConfig, loadEnv } from "vite"
 import { viteStaticCopy } from "vite-plugin-static-copy"
-import zipPack from "vite-plugin-zip-pack"
-
-const pluginInfo = require("./plugin.json")
-
-export default defineConfig(({
-  mode,
-}) => {
-
-  console.log('mode=>', mode)
-  const env = loadEnv(mode, process.cwd())
-  const {
-    VITE_SIYUAN_WORKSPACE_PATH,
-  } = env
-  console.log('env=>', env)
+import livereload from "rollup-plugin-livereload"
+import { svelte } from "@sveltejs/vite-plugin-svelte"
+import zipPack from "vite-plugin-zip-pack";
+import fg from 'fast-glob';
+import fs from 'fs';
+import { execSync } from 'child_process';
 
 
-  const siyuanWorkspacePath = VITE_SIYUAN_WORKSPACE_PATH
-  let devDistDir = './dev'
-  if (!siyuanWorkspacePath) {
-    console.log("\nSiyuan workspace path is not set.")
-  } else {
-    console.log(`\nSiyuan workspace path is set:\n${siyuanWorkspacePath}`)
-    devDistDir = `${siyuanWorkspacePath}/data/plugins/${pluginInfo.name}`
-  }
-  console.log(`\nPlugin will build to:\n${devDistDir}`)
+const env = process.env;
+const isSrcmap = env.VITE_SOURCEMAP === 'inline';
+const isDev = env.NODE_ENV === 'development';
 
-  const args = minimist(process.argv.slice(2))
-  const isWatch = args.watch || args.w || false
-  const distDir = isWatch ? devDistDir : "./dist"
+const outputDir = isDev ? "dev" : "dist";
 
-  console.log()
-  console.log("isWatch=>", isWatch)
-  console.log("distDir=>", distDir)
+console.log("isDev=>", isDev);
+console.log("isSrcmap=>", isSrcmap);
+console.log("outputDir=>", outputDir);
 
-  return {
+export default defineConfig({
     resolve: {
-      alias: {
-        "@": resolve(__dirname, "src"),
-      },
+        alias: {
+            "@": resolve(__dirname, "src"),
+        }
     },
 
     plugins: [
-      viteStaticCopy({
-        targets: [
-          {
-            src: "./README*.md",
-            dest: "./",
-          },
-          {
-            src: "./icon.png",
-            dest: "./",
-          },
-          {
-            src: "./preview.png",
-            dest: "./",
-          },
-          {
-            src: "./plugin.json",
-            dest: "./",
-          },
-          {
-            src: "./src/i18n/**",
-            dest: "./i18n/",
-          },
-          {
-            src: "./mindmap-embed/index.html",
-            dest: "./mindmap-embed/",
-          },
-          {
-            src: "./mindmap-embed/dist/**",
-            dest: "./mindmap-embed/dist/",
-          }
-        ],
-      }),
+        svelte(),
+
+        viteStaticCopy({
+            targets: [
+                { src: "./README*.md", dest: "./" },
+                { src: "./plugin.json", dest: "./" },
+                { src: "./preview.png", dest: "./" },
+                { src: "./icon.png", dest: "./" },
+                { src: "./assets/*", dest: "./assets/" },
+                { src: "./i18n/*", dest: "./i18n/" },
+            ],
+        }),
+
+        // Auto copy to SiYuan plugins directory in dev mode
+        ...(isDev ? [
+            {
+                name: 'auto-copy-to-siyuan',
+                writeBundle() {
+                    try {
+                        // Run the copy script after build
+                        execSync('node --no-warnings ./scripts/make_dev_copy.js', {
+                            stdio: 'inherit',
+                            cwd: process.cwd()
+                        });
+                    } catch (error) {
+                        console.warn('Auto copy to SiYuan failed:', error.message);
+                        console.warn('You can manually run: pnpm run make-link-win');
+                    }
+                }
+            }
+        ] : []),
+
     ],
 
-    // https://github.com/vitejs/vite/issues/1930
-    // https://vitejs.dev/guide/env-and-mode.html#env-files
-    // https://github.com/vitejs/vite/discussions/3058#discussioncomment-2115319
-    // 在这里自定义变量
     define: {
-      "process.env.DEV_MODE": `"${isWatch}"`,
-      "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV),
+        "process.env.DEV_MODE": JSON.stringify(isDev),
+        "process.env.NODE_ENV": JSON.stringify(env.NODE_ENV)
     },
 
     build: {
-      // 输出路径
-      outDir: distDir,
-      emptyOutDir: !isWatch,
+        outDir: outputDir,
+        // Keep existing files in output directory for incremental builds
+        emptyOutDir: false,
+        minify: true,
+        sourcemap: isSrcmap ? 'inline' : false,
 
-      // 构建后是否生成 source map 文件
-      sourcemap: false,
-
-      // 设置为 false 可以禁用最小化混淆
-      // 或是用来指定是应用哪种混淆器
-      // boolean | 'terser' | 'esbuild'
-      // 不压缩，用于调试
-      minify: !isWatch,
-
-      lib: {
-        // Could also be a dictionary or array of multiple entry points
-        entry: resolve(__dirname, "src/index.ts"),
-        // the proper extensions will be added
-        fileName: "index",
-        formats: ["cjs"],
-      },
-      rollupOptions: {
-        plugins: [
-          ...(isWatch
-            ? [
-                livereload(devDistDir),
-                {
-                  // 监听静态资源文件
-                  name: "watch-external",
-                  async buildStart() {
-                    const files = await fg([
-                      "src/i18n/*.json",
-                      "./README*.md",
-                      "./plugin.json",
-                    ])
-                    for (const file of files) {
-                      this.addWatchFile(file)
-                    }
-                  },
-                },
-              ]
-            : [
-                zipPack({
-                  inDir: "./dist",
-                  outDir: "./",
-                  outFileName: "package.zip",
-                }),
-              ]),
-        ],
-
-        // make sure to externalize deps that shouldn't be bundled
-        // into your library
-        external: ["siyuan", "process"],
-
-        output: {
-          entryFileNames: "[name].js",
-          assetFileNames: (assetInfo) => {
-            if (assetInfo.name === "style.css") {
-              return "index.css"
-            }
-            return assetInfo.name
-          },
+        lib: {
+            entry: resolve(__dirname, "src/index.ts"),
+            fileName: "index",
+            formats: ["cjs"],
         },
-      },
-    },
-  }
-})
+        rollupOptions: {
+            plugins: [
+                ...(isDev ? [
+                    {
+                        name: 'watch-external',
+                        async buildStart() {
+                            const files = await fg([
+                                './i18n/**',
+                                './README*.md',
+                                './plugin.json'
+                            ]);
+                            for (let file of files) {
+                                this.addWatchFile(file);
+                            }
+                        }
+                    }
+                ] : [
+                    // Clean up unnecessary files under dist dir
+                    cleanupDistFiles({
+                        patterns: ['i18n/*.yaml', 'i18n/*.md'],
+                        distDir: outputDir
+                    }),
+                    zipPack({
+                        inDir: './dist',
+                        outDir: './',
+                        outFileName: 'package.zip'
+                    })
+                ])
+            ],
+
+            external: ["siyuan", "process"],
+
+            output: {
+                entryFileNames: "[name].js",
+                assetFileNames: (assetInfo) => {
+                    if (assetInfo.name === "style.css") {
+                        return "index.css"
+                    }
+                    return assetInfo.name
+                },
+            },
+        },
+    }
+});
+
+
+/**
+ * Clean up some dist files after compiled
+ * @author frostime
+ * @param options:
+ * @returns 
+ */
+function cleanupDistFiles(options: { patterns: string[], distDir: string }) {
+    const {
+        patterns,
+        distDir
+    } = options;
+
+    return {
+        name: 'rollup-plugin-cleanup',
+        enforce: 'post',
+        writeBundle: {
+            sequential: true,
+            order: 'post' as 'post',
+            async handler() {
+                const fg = await import('fast-glob');
+                const fs = await import('fs');
+                // const path = await import('path');
+
+                // 使用 glob 语法，确保能匹配到文件
+                const distPatterns = patterns.map(pat => `${distDir}/${pat}`);
+                console.debug('Cleanup searching patterns:', distPatterns);
+
+                const files = await fg.default(distPatterns, {
+                    dot: true,
+                    absolute: true,
+                    onlyFiles: false
+                });
+
+                // console.info('Files to be cleaned up:', files);
+
+                for (const file of files) {
+                    try {
+                        if (fs.default.existsSync(file)) {
+                            const stat = fs.default.statSync(file);
+                            if (stat.isDirectory()) {
+                                fs.default.rmSync(file, { recursive: true });
+                            } else {
+                                fs.default.unlinkSync(file);
+                            }
+                            console.log(`Cleaned up: ${file}`);
+                        }
+                    } catch (error) {
+                        console.error(`Failed to clean up ${file}:`, error);
+                    }
+                }
+            }
+        }
+    };
+}
