@@ -1,10 +1,13 @@
-/**
- * Copyright (c) 2023 frostime. All rights reserved.
+/*
+ * Copyright (c) 2024 by frostime. All Rights Reserved.
+ * @Author       : frostime
+ * @Date         : 2023-08-15 10:28:10
+ * @FilePath     : /src/types/index.d.ts
+ * @LastEditTime : 2024-06-08 20:50:53
+ * @Description  : Frequently used data structures in SiYuan
  */
 
-/**
- * Frequently used data structures in SiYuan
- */
+
 type DocumentId = string;
 type BlockId = string;
 type NotebookId = string;
@@ -12,154 +15,92 @@ type PreviousID = BlockId;
 type ParentID = BlockId | DocumentId;
 
 type Notebook = {
-  id: NotebookId;
-  name: string;
-  icon: string;
-  sort: number;
-  closed: boolean;
-};
+    id: NotebookId;
+    name: string;
+    icon: string;
+    sort: number;
+    closed: boolean;
+}
 
 type NotebookConf = {
-  name: string;
-  closed: boolean;
-  refCreateSavePath: string;
-  createDocNameTemplate: string;
-  dailyNoteSavePath: string;
-  dailyNoteTemplatePath: string;
-};
+    name: string;
+    closed: boolean;
+    refCreateSavePath: string;
+    createDocNameTemplate: string;
+    dailyNoteSavePath: string;
+    dailyNoteTemplatePath: string;
+}
 
-type BlockType =
-  | "d"
-  | "s"
-  | "h"
-  | "t"
-  | "i"
-  | "p"
-  | "f"
-  | "audio"
-  | "video"
-  | "other";
+type BlockType = 
+    | 'd'
+    | 'p'
+    | 'query_embed'
+    | 'l'
+    | 'i'
+    | 'h'
+    | 'iframe'
+    | 'tb'
+    | 'b'
+    | 's'
+    | 'c'
+    | 'widget'
+    | 't'
+    | 'html'
+    | 'm'
+    | 'av'
+    | 'audio';
 
-type BlockSubType =
-  | "d1"
-  | "d2"
-  | "s1"
-  | "s2"
-  | "s3"
-  | "t1"
-  | "t2"
-  | "h1"
-  | "h2"
-  | "h3"
-  | "h4"
-  | "h5"
-  | "h6"
-  | "table"
-  | "task"
-  | "toggle"
-  | "latex"
-  | "quote"
-  | "html"
-  | "code"
-  | "footnote"
-  | "cite"
-  | "collection"
-  | "bookmark"
-  | "attachment"
-  | "comment"
-  | "mindmap"
-  | "spreadsheet"
-  | "calendar"
-  | "image"
-  | "audio"
-  | "video"
-  | "other";
+
+type BlockSubType = "d1" | "d2" | "s1" | "s2" | "s3" | "t1" | "t2" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "table" | "task" | "toggle" | "latex" | "quote" | "html" | "code" | "footnote" | "cite" | "collection" | "bookmark" | "attachment" | "comment" | "mindmap" | "spreadsheet" | "calendar" | "image" | "audio" | "video" | "other";
 
 type Block = {
-  id: BlockId;
-  parent_id?: BlockId;
-  root_id: DocumentId;
-  hash: string;
-  box: string;
-  path: string;
-  hpath: string;
-  name: string;
-  alias: string;
-  memo: string;
-  tag: string;
-  content: string;
-  fcontent?: string;
-  markdown: string;
-  length: number;
-  type: BlockType;
-  subtype: BlockSubType;
-  /** string of { [key: string]: string }
-   * For instance: "{: custom-type=\"query-code\" id=\"20230613234017-zkw3pr0\" updated=\"20230613234509\"}"
-   */
-  ial?: string;
-  sort: number;
-  created: string;
-  updated: string;
-};
+    id: BlockId;
+    parent_id?: BlockId;
+    root_id: DocumentId;
+    hash: string;
+    box: string;
+    path: string;
+    hpath: string;
+    name: string;
+    alias: string;
+    memo: string;
+    tag: string;
+    content: string;
+    fcontent?: string;
+    markdown: string;
+    length: number;
+    type: BlockType;
+    subtype: BlockSubType;
+    /** string of { [key: string]: string } 
+     * For instance: "{: custom-type=\"query-code\" id=\"20230613234017-zkw3pr0\" updated=\"20230613234509\"}" 
+     */
+    ial?: string;
+    sort: number;
+    created: string;
+    updated: string;
+}
 
 type doOperation = {
-  action: string;
-  data: string;
-  id: BlockId;
-  parentID: BlockId | DocumentId;
-  previousID: BlockId;
-  retData: null;
-};
+    action: string;
+    data: string;
+    id: BlockId;
+    parentID: BlockId | DocumentId;
+    previousID: BlockId;
+    retData: null;
+}
 
 interface Window {
-  siyuan: {
-    notebooks: any;
-    menus: {
-      menu: Menu;
+    siyuan: {
+        config: any;
+        notebooks: any;
+        menus: any;
+        dialogs: any;
+        blockPanels: any;
+        storage: any;
+        user: any;
+        ws: any;
+        languages: any;
+        emojis: any;
     };
-    dialogs: any;
-    blockPanels: any;
-    storage: any;
-    user: any;
-    ws: any;
-    languages: any;
-    config: {
-      appearance: {
-        mode: number;
-      };
-      lang: string;
-      keymap: any;
-    }
-  };
-  _sy_plugin_sample: {
-    [key: string]: any;
-  };
-};
-
-
-enum SyFrontendTypes {
-  // 桌面端
-  'desktop' = 'desktop',
-  'desktop-window' = 'desktop-window',
-  // 移动端
-  'mobile' = 'mobile',
-  // 浏览器 - 桌面端
-  'browser-desktop' = 'browser-desktop',
-  // 浏览器 - 移动端
-  'browser-mobile' = 'browser-mobile',
-}
-
-interface MindmapImageInfo {
-  imageURL: string;
-  data: string;
-  format: "svg" | "png";
-  blockID?: string;
-}
-
-interface SettingItem {
-    title?: string;
-    description?: string;
-    direction?: "row" | "column";
-    actionElement?: HTMLElement;
-    createActionElement?: () => HTMLElement;
+    Lute: any;
 }
